refactor: move `FixedLoggerTime` into its own test helper file
### Security

- Use a runtime-created log directory for the demo logger destination.
